% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facedata_mm}
\alias{facedata_mm}
\title{Face anthropometry interval data}
\format{
An \code{interval_data} object with 27 observations and 6 variables:
\describe{
\item{centers}{27 x 6 matrix of interval midpoints.}
\item{radii}{27 x 6 matrix of interval half-widths.}
\item{labels}{Factor with 9 levels (3-letter person identifiers).}
}
}
\source{
Billard, L. and Diday, E. (2006).
\emph{Symbolic Data Analysis: Conceptual Statistics and Data Mining}.
Wiley.
}
\usage{
facedata_mm
}
\description{
Interval-valued anthropometric face measurements of 27 individuals
across 6 variables: AD, BC, AH, DH, EH, GH (distances between
facial landmarks).
}
\keyword{datasets}
