% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval-data.R
\name{interval_data_from_array}
\alias{interval_data_from_array}
\title{Create interval_data from a 3D min-max array}
\usage{
interval_data_from_array(arr, labels = NULL)
}
\arguments{
\item{arr}{A 3D numeric array of dimensions (n, p, 2).}

\item{labels}{An optional factor or character vector of class labels.}
}
\value{
An \code{interval_data} object.
}
\description{
Converts a 3D array (n x p x 2) where \code{[,,1]} is the lower bound
and \code{[,,2]} is the upper bound into an \code{interval_data} object.
}
\examples{
arr <- array(c(1, 2, 3, 4, 5, 6, 2, 4, 5, 6, 8, 9),
             dim = c(3, 2, 2))
x <- interval_data_from_array(arr)
}
