% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr-methods.R
\name{umap_config_default}
\alias{umap_config_default}
\title{Default UMAP configuration for interval UMAP}
\usage{
umap_config_default()
}
\value{
A named list of UMAP configuration parameters.
}
\description{
Returns a list of UMAP hyperparameters with default settings.
}
\examples{
cfg <- umap_config_default()
}
