% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QAIDR-package.R
\docType{package}
\name{QAIDR-package}
\alias{QAIDR}
\alias{QAIDR-package}
\title{QAIDR: Quality Assessment for Interval-Based Dimensionality Reduction}
\description{
Provides quality and behavior indices for evaluating interval-based dimensionality reduction (DR) methods using co-ranking matrices. Supports multiple interval distance metrics (Euclidean, Hausdorff, Ichino-Yaguchi, L2-Wasserstein) and assessment criteria (Trustworthiness & Continuity, MRRE, LCMC). Includes wrappers for symbolic PCA methods, interval MDS, and interval UMAP, as well as permutation tests for statistical significance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hanmingwu1103/QAIDR}
  \item Report bugs at \url{https://github.com/hanmingwu1103/QAIDR/issues}
}

}
\author{
\strong{Maintainer}: Han-Ming Wu \email{hmwu@stat.ntpu.edu.tw} (\href{https://orcid.org/0000-0002-0656-374X}{ORCID})

}
\keyword{internal}
